Puzzle
Given size N, randomly generate a sliding puzzle that can be solved by
the user. Add hint button that will tell the user the next best step if
they get stuck. If the user decides to give up, allow them to see the
sequence of moves needed to solve the puzzle.